requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.42.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.2
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
comtypes==1.2.0
configparser==5.3.0
debugpy==1.6.3
decorator==5.1.1
deprecation==2.1.0
docutils==0.20.1
Eel==0.16.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fastapi==0.88.0
ffpass==0.5.0
filelock==3.10.4
Flask==2.2.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
frozenlist==1.3.3
ftfy==6.1.1
future==0.18.2
gevent==23.7.0
gevent-eventemitter==2.1
gevent-websocket==0.10.1
googletrans==3.0.0
GPUtil==1.4.0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.5.0
httpx==0.13.3
huggingface-hub==0.13.3
hyperframe==5.2.0
idna==2.10
img2pdf==0.4.4
ipykernel==6.15.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
keyboard==0.13.5
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
log-symbols==0.0.14
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mouse==0.7.1
multidict==6.0.4
mutagen==1.47.0
nest-asyncio==1.5.5
Nuitka==2.3.4
numpy==1.23.5
openai==0.27.2
opencv-python==4.6.0.66
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
pikepdf==6.2.1
Pillow==9.3.0
playsound==1.2.2
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pycaw==20230407
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.18.0
pydantic==1.10.2
pydub==0.25.1
pygame==2.5.0
pyglet==2.0.9
Pygments==2.13.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2022.12
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.9.3.3
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
PySide6==6.4.3
PySide6-Addons==6.4.3
PySide6-Essentials==6.4.3
pyTelegramBotAPI==4.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
pywin32==304
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq==23.2.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
questionary==1.10.0
regex==2023.3.23
requests==2.28.1
rfc3986==1.5.0
sentencepiece==0.1.97
shiboken6==6.4.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
spinners==0.0.24
SQLAlchemy==2.0.28
SQLAlchemy-serializer==1.4.1
stack-data==0.5.0
starlette==0.22.0
steam==1.4.4
tabulate==0.9.0
termcolor==2.2.0
tokenizers==0.13.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
typing_extensions==4.10.0
urllib3==1.26.12
uvicorn==0.20.0
vdf==3.4
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
whichcraft==0.6.1
WTForms==3.1.2
xonsh==0.13.4
yarl==1.8.2
zope.event==5.0
zope.interface==6.0
zstandard==0.22.0
